06.Server-integration/02.Preauthorizing-devices: remove Device Admission refs + other issues
removing refs to deviceadm is one thing, but this section had
some other issues ('Make sure there are no existing authentication
sets for your device'):
- both /devices endpoints are paged - no mention of it, so the user
  might miss entries past page 1
- then there's some redundancy - if we recommend cleanup via
  DELETE devauth/devices/:id - that's the decommissioning EP, so it's
  enough to destroy all of the device's data in the system,
  no need to recommend DELETE admission/devices/:id in addition

deviceadm is not there any more, so cleaning up deviceauth is all
there's left. in the meantime, we got a CLI command to do just that,
mentioned in the 'Troubleshooting' section. all in all most of
this section is removed and the user is redirected to there.