use 'jq' utility to decoded and pretty-print JSON
Instead of "python -m json.tool".

It is commonly used together with 'curl' and JSON, but it is also
required to decode certain values in responses, e.g an artifact
URI will contain raw values that need to be decoded by a JSON
parser.

Some more background info can be found in MEN-2082.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>