Avoid suggesting state script that can break migration.
Clarify that it is ArtifactInstall_Leave that should be used for
migrations, not ArtifactReboot_Enter. This is important because if a
spontaneous reboot happens while inside ArtifactReboot_Enter, it will
be ignored, and hence the new partition will be entered with an
incomplete migration in place. With ArtifactInstall_Leave it will
instead cause a rollback, triggering a potential ArtifactFailure
script which can undo the half done migration.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 49bf95e03418a1eab2e14bfa9e3c1e4e2ba0fc24)