Updated Mender artifact server IP command
"echo" with a Mender required line will copy only mender related line but, does not guarantee other lines which might be required for the platform.
Hence, updated the command similar to the "sdimg" way (i.e. cat followed by sed and append the line at the end of the existing file) to make it more generic.