Improve instructions for mender-artifact build: version and deps
We want the instructions to point to the latest released version. This
is important now that for the 3.0.0b release the code is not merged into
master, so without the checkout we cannot build mender-artifact suitable
for Update Modules

Also, document de LZMA depdendency with specific instructions for Mac OS
users, as we do not provice a native binary for them (yet).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit da53725079c41de5f0e69f2ce50f65e105a3b233)