fix link to patch in troubleshooting section
The actual link in markdown is valid but it seems to be mangled, probably
due to the http arguments in the call and resulting in a 404.

Update the patch link to point to github hosted repository.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>
(cherry picked from commit 6d21f93573f34b02bbc9c50b688c9291299c5477)