state-script documentation cleanup
Most significantly:

Moved the bulk of the description (updated) of the variables listed below to the Mender
client configuration section, mostly to avoid duplication of default settings and to avoid
repeating the information. This is after all an more advanced topic of the Mender client,
and does not need be covered in one single page.

- StateScriptRetryIntervalSeconds
- StateScriptRetryTimeoutSeconds
- StateScriptTimeoutSeconds

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>