Fix dead internal link
See: https://github.com/remarkjs/remark-validate-links/issues/54 for why this
was not caught by the internal link checker.

This was a curious bug, and a result of a bug in my link migration script it
seems. bash '${VAR##.*#}' anyone? Seems it messed up the suffix placement :)

I also did a regex search for similar cases, and found none other than this one.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>