Add usability improvements like help and verbose
The standard setup was a little verbose, printing all passive voice occurences,
which was nice as a helper tool. But now, the tool is to be run in CI, and we're
only interested in failures.

Therefore, it now prints only passive voice in files which fails the passive
voice test. The full print can still be enabled through the '--verbose' flag.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>