Add a script for running the passive-voice script on all doc files
This script runs './scripts/passive.sh' on all documentation files present in
the directory, but only after filtering out the matches in the
'.passive-blacklist' file in the root directory. This blacklist contains grep
expressions which are run on all the doc files present. For now, it is a catch
all for all directories except the first two, but will eventually be pruned
down, as passive voice is weeded out.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>