MEN-5391: autoversion: Add logic to query either git or docker version
With our latest release_tool we cannot anymore use the "trick" of
querying docker versions to pure git components (like mender-artifact,
mender, etc) and expect it to fallback to git version and work.

By default the tool will now use the git versions, unless an special
string is used for in the AUTOVERSION expression to trigger the switch
to docker version. See updated README.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>