autoversion.py: Switch to git versions
With the recent changes on release tool, we cannot anymore query
"docker" versions for repositories with no Docker deliverable.

The only reason we were using docker versions was for this unique
reference to a Docker image. Switch to git versions which works nicely
with all client repositories and modify the version string on Docker
images to mender-<integ-version>

This is a simpler fix than b37a34f.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>