fix: corrects misleading configuration for s3 buckets
Documentation is misleading. Including the bucket name of an S3 bucket in both AWS_URI and AWS_BUCKET causes the deployments container to look for “https://<AWS_BUCKET>.<name-of-your-bucket>.s3.<your-aws-region>.amazonaws.com” which is not valid. It is prepending the AWS_BUCKET to the front of the url.

Simply including the name in the AWS_URI is sufficient.

Signed-off-by: Mathew Akin <mathew@byos.io>