fix: Remove the reference to Debian buster
The removal led to a lot of changes. With doing it this way we keep the
resources for installing the client in the downloads section, which makes sure
we don't drift away from the steps required in different parts of the docs.

Ticket: MEN-5907
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit 5013258e1ac8664d7f7496db9bcf3f874e0e6bfb)
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>