ci: Hard-code mender-ci-workflows-version to 1.0.0
As this component is out of the release_tool, the ci test has no way to
know which is the latest version (for now).

Amends commit 72079d479137940b6afcecf35c2f8271585dc335 and it will be
reworked together with the original commit once we have a design for
where to store the latest version of independent components.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>