fix: document Raspberry Pi Imager bug
The Raspberry Pi Imager offers options to pre-customize the
image right during flashing to the SD card. This injects
a firstrun.sh script, along with a kernel command line that
causes an infinite boot loop.
This needs to be documented in the onboarding section as well
as the troubleshoot section.

Changelog: Title
Ticket: None

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>