Update docs.md
There was a bit of messy typing here so I thought I'd suggest a fix. Looking at it though I think it is better to not include a link to the place you don't want them to go until they read the following text so I removed the reference and link to the next section: Kubernetes.

Signed-off-by: Craig Comstock <craig.comstock@northern.tech>