chore: example on how to set the mtls proxy with kubernetes
This focuses only on the server side, giving an example on how to set
it.
The client form the docker can be used but it isn't the focus

This focuses only on the server side, giving an example on how to set
it.
The client form the docker can be used but it isn't the focus.

Signed-off-by: Alan <alan.martinovic@northern.tech>