[chore] Refresh of the docker compose docs
Port forwarding expects the evaluator to both install mender-cli and figure out a device ID.
Those are extra steps that risk confusion and preventing the reader to
reach a working example for the docker compose update.
I've instead just replaced that with the Remote terminal for all client
shell interactions.

The versioning was hardcoded and the confusions between composition
version vs container version were left to the user to wrestle with.
Added a section to clarify this and adjusted the snippets.

For this case the benefits aren't spectacular, but that is heavy content
specific. I wanted to show there is indeed an improvement and give a
quick snippet for the evaluator to test it. If they have the question
"Is delta beneficial for this composition I want to use?" they can
easily use the same approach.

Ticket: None
Changelog: None

Signed-off-by: Alan <alan.martinovic@northern.tech>