Import READMEs verbatim from the orchestrator repo.
This commit only exists to make the review of the next commit easier,
so one can easily see what has been changed from the original when
basing the docs on these READMEs.

A few portions that dealt with compiling the source code has been
taken out, since the repository is closed source.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>