chore: Repurpose `autoversion.py` script to fit release split
The releases won't anymore follow a monolithic process, and there won't
be one single place were to gather release versions from.

That sounds like a drama, but it is not! Hear me out:

We are moving to a flat structure in the docs content (iow no parallel
content sites) which means that we will have the current versions in
`mender-docs/master`, and then only update a given component when
released. For example a release of one tool will do a single call like:
`autoversion.py --update --component mender-cli --version 1.2.3` (which,
by the way, is super easy to integrate into a post tag action from CI).

All right, so what is this commit doing:
* Remove dependency of release_tool and integration repo checkout
* Remove parsing of releases.json file
* Remove support for the "lts" replacing magic. This is a downgrade if
  you ask me, as it will force human intervention. In any case that page
  is under heavy redesign by the Product Owner so the value of the
  automation is to be seen.
* A nice help to fail if the intended component was not found (a user
  types `mender-client` instead of `mender`, for example).

Kudos for reading all the way to here. Thank you.

Ticket: QA-794

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>