docs: Replace libssl1.1 `mender-artifact` with deb packages
OpenSSL 1.1 has been replaced in latest Ubuntu and Debian releases and,
most importantly, Ubuntu does not provide an openssl1.1 compatibility
package, which runs our pre-compiled binary unusable there.

Instead of having two binaries, one per each OpenSSL version, we link
now directly into the deb packages as the preferred method to install
the tool on Linux.

On Mac the support has not changed, but we add the architecture to make
it explicit where the binary is expected to work.

As a side effect of using deb packages, we cannot offer anymore a
reliable link to the pre-release version, so we remove these from the
table. The master versions are still accessible via APT experimental
repositories for the advance users.

Ticket: MEN-7767

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>