docs: Added note about how to achieve time sync before mender services start
It's not uncommon for this to be an issue for users. Providing a bit more
specific guidance on how to ensure correct time prior to mender services
starting could help some users avoid running into issues where because of large
polling intervals for production rejections related to time sync can cause long
delays in establishing communication with the server.

Changelog: None
Ticket: CE-564
Signed-off-by: Nick Anderson <nick@cmdln.org>