chore: Strictly use Device System in Gateway getting started
A *Device System* is a concept from Mender Gateway that predates
the *System devices* introduced with the Orchestator. They are
similar concepts, but at different levels and we need to make
sure users understand these are different things.

Let's strictly use *Device System* for Gateway's device systems
and add a note about Orchestrator System devices with a link to
the Orchestrator docs.

Ticket: MEN-8409
Changelog: none
Signed-off-by: Vratislav Podzimek <vratislav.podzimek+auto-signed@northern.tech>